Fix the barcode reader.
The problem was that the RWH guys represented a bitmap as a ((0, 0), (w-1, h-1))
Array. Arrays are stored "from left to right", which means RWH's bitmaps were
column-major. PGMs are sane and thus store images row-major. I modified the code
to be row-major throughout, which means bitmaps are now ((0, 0), (h-1, w-1)) arrays.